Remove internal interfaces
